Adding disk support and writing a base class for heartbeats. Needs work

+++ src/WatchServer.Core/Services/RAMHeartbeatService.cs
@@ -21,7 +21,7 @@ namespace WatchServer.Core.Services
 
         public void WriteMetric()
         {
-            _reportingService.WriteMetric(MetricCode.PercentRAM, ((int)_ramCounter.NextValue()) / _totalRam);
+            _reportingService.WriteMetric(MetricCode.PercentRAM, _ramCounter.NextValue() / _totalRam);
         }
 
         public void StartCollecting()
